refactor: moves inline fields to own dir
